+++ metadata-ingestion/src/datahub/ingestion/source/sql/bigquery.py
@@ -1,6 +1,6 @@
 import collections
-import datetime
 import dataclasses
+import datetime
 import functools
 import json
 import logging
@@ -27,8 +27,8 @@ from datahub.configuration import ConfigModel
 from datahub.configuration.common import ConfigurationError
 from datahub.configuration.time_window_config import BaseTimeWindowConfig
 from datahub.emitter import mce_builder
-from datahub.emitter.mce_builder import PlatformKey, gen_containers
 from datahub.emitter.mcp import MetadataChangeProposalWrapper
+from datahub.emitter.mcp_builder import PlatformKey, gen_containers
 from datahub.ingestion.api.workunit import MetadataWorkUnit
 from datahub.ingestion.source.sql.sql_common import (
     SQLAlchemyConfig,
@@ -764,10 +764,15 @@ WHERE
         )
 
     def gen_database_containers(self, database: str) -> Iterable[MetadataWorkUnit]:
+        domain_urn = self._gen_domain_urn(database)
+
         database_container_key = self.gen_database_key(database)
 
         container_workunits = gen_containers(
-            database_container_key, database, ["Project"]
+            container_key=database_container_key,
+            name=database,
+            sub_types=["Project"],
+            domain_urn=domain_urn,
         )
 
         for wu in container_workunits:
