+++ metadata-ingestion/src/datahub/ingestion/source/state/stateful_ingestion_base.py
@@ -30,7 +30,7 @@ class StatefulIngestionConfig(ConfigModel):
     """
 
     enabled: bool = False
-    max_checkpoint_state_size: int = 2**24  # 16MB
+    max_checkpoint_state_size: int = 2 ** 24  # 16MB
     state_provider: Optional[DynamicTypedConfig] = DynamicTypedConfig(
         type="datahub", config=DatahubIngestionStateProviderConfig()
     )
