+++ metadata-ingestion/tests/unit/test_mce_builder.py
@@ -21,21 +21,3 @@ def test_can_add_aspect():
     assert builder.can_add_aspect(dataset_mce, DatasetPropertiesClass)
     assert builder.can_add_aspect(dataset_mce, OwnershipClass)
     assert not builder.can_add_aspect(dataset_mce, DataFlowInfoClass)
-
-
-def test_guid_generator():
-    key = builder.SchemaKey(
-        database="test", schema="Test", platform="mysql", instance="PROD"
-    )
-
-    guid = key.guid()
-    assert guid == "06a1f87f99e1d82efa3da13637913c87"
-
-
-def test_guid_generator_with_empty_instance():
-    key = builder.SchemaKey(
-        database="test", schema="Test", platform="mysql", instance=None
-    )
-
-    guid = key.guid()
-    assert guid == "0ce13865e9e414406a895612787d9ae6"
